+++ R/elemental_index.R
@@ -37,7 +37,8 @@ different_length <- function(...) {
 #' Make elemental price indexes
 #'
 #' Compute period-over-period (chainable) or fixed-base (direct) elemental
-#' price indexes, with optional percent-change contributions.
+#' price indexes, with optional percent-change contributions for each
+#' product.
 #'
 #' When supplied with a numeric vector, `elemental_index()` is a simple
 #' wrapper that applies
@@ -52,7 +53,9 @@ different_length <- function(...) {
 #' elemental indexes, and chapter 5 of Balk (2008).
 #'
 #' The default method simply coerces `x` to a numeric vector prior to
-#' calling the method above.
+#' calling the method above. The data frame method provides a formula interface
+#' to specify columns of price relatives, time periods, and elemental
+#' aggregates and call the method above.
 #'
 #' Names for `x` are used as product names when calculating percent-change
 #' contributions. Product names should be unique within each time period, and,
@@ -83,8 +86,8 @@ different_length <- function(...) {
 #' are calculated.
 #'
 #' @param x Period-over-period or fixed-base price relatives. Currently there
-#' is only a method for numeric vectors; these can be made with
-#' [price_relative()].
+#' are methods for numeric vectors (which can be made with
+#' [price_relative()]) and data frames.
 #' @param period A factor, or something that can be coerced into one, giving
 #' the time period associated with each price relative in `x`. The
 #' ordering of time periods follows of the levels of `period`, to agree
@@ -110,6 +113,9 @@ different_length <- function(...) {
 #' a Paasche index). Other values are possible; see
 #' [gpindex::generalized_mean()] for details.
 #' @param ... Further arguments passed to or used by methods.
+#' @param formula A two-part formula with price relatives on the left-hand
+#' side, and time periods and elemental aggregates (in that order) on the
+#' right-hand side.
 #'
 #' @returns
 #' A price index that inherits from [`piar_index`]. If
@@ -154,7 +160,7 @@ different_length <- function(...) {
 #'
 #' # Calculate Jevons elemental indexes
 #'
-#' with(prices, elemental_index(rel, period, ea))
+#' elemental_index(prices, rel ~ period + ea)
 #'
 #' # Same as using lm() or tapply()
 #'
@@ -174,13 +180,11 @@ different_length <- function(...) {
 #' # Calculate a CSWD index (same as the Jevons in this example)
 #' # as an arithmetic index by using the appropriate weights
 #'
-#' with(
+#' elemental_index(
 #'   prices,
-#'   elemental_index(
-#'     rel, period, ea,
-#'     fw(rel, group = interaction(period, ea)),
-#'     r = 1
-#'   )
+#'   rel ~ period + ea,
+#'   fw(rel, group = interaction(period, ea)),
+#'   r = 1
 #' )
 #'
 #' @export
@@ -192,6 +196,35 @@ elemental_index <- function(x, ...) {
 #' @export
 elemental_index.default <- function(x, ...) {
   elemental_index(as.numeric(x), ...)
+}
+
+#' @rdname elemental_index
+#' @export
+elemental_index.data.frame <- function(x,
+                                       formula,
+                                       weights = NULL, ...,
+                                       chainable = TRUE,
+                                       na.rm = FALSE,
+                                       contrib = FALSE,
+                                       r = 0) {
+  if (length(formula) != 3L) {
+    stop("'formula' must have a left-hand and right-hand side")
+  }
+  fterms <- stats::terms(formula, data = x)
+  if (length(attr(fterms, "term.labels")) != 2L) {
+    stop("right-hand side of 'formula' must have exactly two terms")
+  }
+  weights <- eval(substitute(weights), x, parent.frame())
+  x <- eval(attr(fterms, "variables"), x, environment(formula))
+  
+  elemental_index(
+    x[[1L]], x[[2L]], x[[3L]],
+    weights = weights,
+    chainable = chainable,
+    na.rm = na.rm,
+    contrib = contrib,
+    r = r
+  )
 }
 
 #' @rdname elemental_index
@@ -217,7 +250,7 @@ elemental_index.numeric <- function(x,
     stop("input vectors must be the same length")
   }
   if (any(x <= 0, na.rm = TRUE) || any(weights <= 0, na.rm = TRUE)) {
-    warning("some elements of 'x or 'weights' are less than or equal to 0")
+    warning("some elements of 'x' or 'weights' are less than or equal to 0")
   }
 
   if (contrib) {